Get a basic demo working on the nucleo board.
This is just a basic demo that is now working on the nucleo
board. This is not intended to be merged into the master branch.
This is confirmed to work.

Interestingly, it looks like the API for uC/OS III is different
between the simulation and board versions. This can be seen in the
OS_CPU_SysTickInit function, which takes an argument for the
Cortex-M4 port, but not for the x86 variant. I suggest looking at
[this commit](https://github.com/lhr-solar/BPS/commit/81faefbc6d5b6d39b83fdbbf7a9a17bf91680249)
on the BPS repo, which solves the problem of determining whether
we are in a simulation at compile time, so that the appropriate calls
can be made.

I also suggest modifying the Makefile to include the flash option in the
root file so that we can still flash to the boards when needed.