CAN Simulation (#29)
Simulates both CAN buses connected to the Controls system. CAN1 is the CAN line connected to the rest of the car's systems. This is simulated simply by creating a line in a file where messages can be sent by the Controls system and will be displayed on the GUI. As of right now, this does not allow us to send messages from the GUI (to mimic sending messages from other systems), but that is not incredibly pertinent to our system's operations at this point. That can be added in the future. CAN2 is the CAN line connected to the motor controller. This is simulated in a similar fashion in addition to the logic for the motor controller. The MotorController.py script expects a certain message ID every 250ms and will go into neutral if it does not receive that message. It also simulates a linear acceleration of 0.5 m/s per 250ms (in either direction depending on the relationship between the desired velocity and current velocity