Add an (untested) compatible driver for communicating with the Gecko.
Problem
=======
We need to be able to communicate with the Gecko board, which we are
using to display information.

Solution
========
This adds a driver which uses USART2 to communicate with the Gecko
board, using the same protocol as what should currently be running
on that board.

Notes
=====
This is untested as of the time of this commit. Please make sure to
test this code before choosing to rely on it.

Future Steps
============
Also, I personally do not like this protocol, so if this needs to be
extended in the future to accomodate other status information beyond
speed, cruise control state, regen state, and CAN error indicators,
then I would suggest changing from a string-based communication to a
bitstream encoding the struct that stores all that information. This
would require changing the code on the Gecko as well.