Displayboard (#36)
* Add an (untested) compatible driver for communicating with the Gecko.

Problem
=======
We need to be able to communicate with the Gecko board, which we are
using to display information.

Solution
========
This adds a driver which uses USART2 to communicate with the Gecko
board, using the same protocol as what should currently be running
on that board.

Notes
=====
This is untested as of the time of this commit. Please make sure to
test this code before choosing to rely on it.

Future Steps
============
Also, I personally do not like this protocol, so if this needs to be
extended in the future to accomodate other status information beyond
speed, cruise control state, regen state, and CAN error indicators,
then I would suggest changing from a string-based communication to a
bitstream encoding the struct that stores all that information. This
would require changing the code on the Gecko as well.

* Added UART enum to account for two UART devices

* Made UART_Read more flexible and updated UART tests

* Modified UART write function to account for more than one UART devices

* Modified Write function of UART

* fixed small double declaration of same variable bugs in write function of UART file

* Fixed UART tests

* Fixed test

* Fixed UART data not being written to csv correctly

* minor change

* Modified documentation of BSP_UART file due to changes in member functions and deleted file BSP_UART2 because it was integrated into BSP_UART

* Removed extraneous imports and defintions, basic display test

* Wrote test for display

* Added the abilty to send UART messages from the GUI

* Removed extraneous arguments, updated test

Co-authored-by: Chase Block <chase@chaseblock.com>
Co-authored-by: Cam0Cow <roie.roiegal@gmail.com>