Modify the UART drivers to use the new unified API.
Problem
=======
The current code does not compile for the STM32F413. This is because
the simulator and STM versions are using different APIs, which is
causing a conflict. The STM is using the older version, which
entails separate files for each UART peripheral.

Solution
========
This commit updates the UART BSP to use a single set of functions,
where the UART that needs to be addressed is passed in via a
parameter from the outside. This has been verified to compile on
both the simulator and stm32f413 targets. It has also been
verified to run as expected on the stm32f413 nucleo board.

Notes
=====
This also makes a minor modification to the Makefile, which greatly
reduces the amount of time that it takes to compile for both targets,
but especially the stm32f413.