Fixes bug with str not having attribute close() in simulator modules (#45)
Problem
======
When a file in the Data folder had not yet been created, each Python simulation module would create that file and then try to close it. The variable that `close()` was being called on was a `str`, so there were type mismatch errors. 

Solution
======
Use a `with` block to automatically close the file and remove all incorrect closes of the `str` that contained the name and path of the CSV file.