Add a GPIO BSP layer.
It appears that we never actually added a GPIO BSP layer for the
stm32f413 target. This introduces this layer. BSP_GPIO_Init now takes an
additional two arguements: the pins that need configuring and whether they
should be writable or not.

This also adds some functions to read or write from a specific pin on a gpio
port so that we can avoid complicated logic elsewhere in the code.

I did this before realizing that Erick was also working on this, so sorry about
that. Just take what you want from this commit.